+++ lib/entity/transaction/wrapper.dart
@@ -9,20 +9,45 @@ import "package:flow/utils/utils.dart";
 class ExtensionsWrapper {
   Transfer? get transfer =>
       data.firstWhereOrNull((element) => element is Transfer) as Transfer?;
+  set transfer(Transfer? newTransfer) =>
+      _overrideSingle(transfer, Transfer.keyName);
+
   Geo? get geo => data.firstWhereOrNull((element) => element is Geo) as Geo?;
+  set geo(Geo? newGeo) => _overrideSingle(newGeo, Geo.keyName);
 
   final List<TransactionExtension> data;
 
   const ExtensionsWrapper(this.data);
   const ExtensionsWrapper.empty() : data = const [];
 
+  ExtensionsWrapper clone() => ExtensionsWrapper([...data]);
+
   /// Returns a new instance with merged
-  ExtensionsWrapper merge(List<TransactionExtension> newData) {
+  ExtensionsWrapper getMerged(List<TransactionExtension> newData) {
     return ExtensionsWrapper([
       ...data.where(
           (currrent) => !newData.any((newExt) => currrent.key == newExt.key)),
       ...newData,
     ]);
+  }
+
+  /// Returns a new instance with overridden
+  ExtensionsWrapper getOverriden(TransactionExtension? newData, String key) {
+    return clone().._overrideSingle(newData, key);
+  }
+
+  void _remove(String key) {
+    data.removeWhere((element) => element.key == key);
+  }
+
+  void _overrideSingle(
+    TransactionExtension? newSingle,
+    String key,
+  ) {
+    _remove(key);
+    if (newSingle != null) {
+      data.add(newSingle);
+    }
   }
 
   static ExtensionsWrapper parse(String? extra) {