+++ lib/main.dart
@@ -147,18 +147,31 @@ class FlowState extends State<Flow> {
     final List<Locale> systemLocales =
         WidgetsBinding.instance.platformDispatcher.locales;
 
-    final String? country = systemLocales
+    final List<Locale> favorableLocales = systemLocales
         .where(
-          (element) => element.countryCode != null,
+          (locale) => FlowLocalizations.supportedLanguages.any(
+              (flowSupportedLocalization) =>
+                  flowSupportedLocalization.languageCode ==
+                  locale.languageCode),
         )
-        .firstOrNull
-        ?.countryCode;
+        .toList();
 
-    final Locale overriddenLocale =
-        LocalPreferences().localeOverride.value ?? _locale;
+    print("-------------------;");
+    systemLocales.forEach(print);
+    print("-------------------;");
+    favorableLocales.forEach(print);
+    print("-------------------;");
 
-    _locale = Locale(
-        overriddenLocale.languageCode, overriddenLocale.countryCode ?? country);
+    final Locale overriddenLocale = LocalPreferences().localeOverride.value ??
+        favorableLocales.firstOrNull ??
+        _locale;
+
+    print("-------------------;");
+    print(overriddenLocale);
+    print("-------------------;");
+
+    _locale =
+        Locale(overriddenLocale.languageCode, overriddenLocale.countryCode);
     Moment.setGlobalLocalization(
       MomentLocalizations.byLocale(overriddenLocale.code) ??
           MomentLocalizations.enUS(),