+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next
+
+* Added an option to choose location on a map
+* Fixed iOS locale
+
 ## Beta 0.6.2
 
 * Fixed upcoming transactions config wasn't getting loaded initially, closes